modify type of variable c

+++ helloworld.c
@@ -5,7 +5,9 @@ int main(  void ) {
 	int a, b;
 	printf( "pls input a and b\n" );
 	scanf( "%d %d", &a, &b );
-	int c = a + b;
-	printf( "sum of a and b = %d\n", sum );
+	//int c = a + b;
+	double c = a + b;
+	//printf( "sum of a and b = %d\n", sum );
+	printf( "sum of a and b = %f\n", sum );
 	return 0;
 }
